Add BytesEncoder interface and update InvokeContractMethod for unified serialization.
